Improve workload screens - #6455 (#6632)
* Make container tabs horizontal.

* Update cronjobs, DaemonSets, Jobs, StatefulSets to use same Workload component.

* Fix 6755 - Wrong sidecar config edit not shown in UI.

* Set Persistent Volume claim capacity as required.
