+++ shell/detail/workload/index.vue
@@ -120,6 +120,10 @@ export default {
       return this.value.type === WORKLOAD_TYPES.CRON_JOB;
     },
 
+    isPod() {
+      return this.value.type === WORKLOAD_TYPES.POD;
+    },
+
     podSchema() {
       return this.$store.getters['cluster/schemaFor'](POD);
     },
@@ -133,13 +137,15 @@ export default {
     },
 
     podTemplateSpec() {
-      const isCronJob = this.value.type === WORKLOAD_TYPES.CRON_JOB;
-
-      if ( isCronJob ) {
+      if ( this.value.type === WORKLOAD_TYPES.CRON_JOB ) {
         return this.value.spec.jobTemplate.spec.template.spec;
-      } else {
-        return this.value.spec?.template?.spec;
       }
+
+      if ( this.value.type === WORKLOAD_TYPES.POD ) {
+        return this.value;
+      }
+
+      return this.value.spec?.template?.spec;
     },
 
     container() {
